# backend/src/main/java/com/app/visite_technique/controller/ClientController.java
package com.app.visite_technique.controller;

import com.app.visite_technique.entity.Client;
import com.app.visite_technique.service.ClientService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;

@RestController
@RequestMapping("/api/clients")
@CrossOrigin("*")
public class ClientController {

    @Autowired
    private ClientService service;

    @GetMapping
    public Page<Client> getAllClients(
            @RequestParam(required = false) String query,
            @PageableDefault(size = 10) Pageable pageable) {
        if (query != null && !query.isEmpty()) {
            return service.search(query, pageable);
        }
        return service.findAll(pageable);
    }

    @GetMapping("/{id}")
    public ResponseEntity<Client> getClientById(@PathVariable Long id) {
        return service.findById(id)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public Client createClient(@RequestBody Client client) {
        return service.save(client);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Client> updateClient(@PathVariable Long id, @RequestBody Client client) {
        return service.findById(id)
                .map(existing -> {
                    client.setId(id);
                    return ResponseEntity.ok(service.save(client));
                })
                .orElse(ResponseEntity.notFound().build());
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteClient(@PathVariable Long id) {
        return service.findById(id)
                .map(client -> {
                    service.delete(id);
                    return ResponseEntity.noContent().<Void>build();
                })
                .orElse(ResponseEntity.notFound().build());
    }
}